impl: make packages and move classes into appropriate package
