--- house/src/main/kotlin/xyz/bukinator/house/model/HouseRepository.kt
@@ -1,4 +1,0 @@
-package xyz.bukinator.house.model
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface HouseRepository : JpaRepository<House, Long>
