+++ house/src/main/kotlin/xyz/bukinator/house/model/converter/StringToListConverter.kt
@@ -1,4 +1,4 @@
-package xyz.bukinator.house.model
+package xyz.bukinator.house.model.converter
 
 import jakarta.persistence.AttributeConverter
 import jakarta.persistence.Converter
