+++ house/src/main/kotlin/xyz/bukinator/house/model/embeddable/Address.kt
@@ -1,4 +1,4 @@
-package xyz.bukinator.house.model
+package xyz.bukinator.house.model.embeddable
 
 import jakarta.persistence.Column
 import jakarta.persistence.Embeddable
