+++ house/src/main/kotlin/xyz/bukinator/house/model/enumerate/HouseStatus.kt
@@ -1,4 +1,4 @@
-package xyz.bukinator.house.model
+package xyz.bukinator.house.model.enumerate
 
 enum class HouseStatus {
     OPEN,
