+++ house/src/main/kotlin/xyz/bukinator/house/repository/HouseRepository.kt
@@ -2,5 +2,6 @@ package xyz.bukinator.house.repository
 
 import org.springframework.data.jpa.repository.JpaRepository
 import xyz.bukinator.house.model.House
+import java.util.UUID
 
-interface HouseRepository : JpaRepository<House, Long>
+interface HouseRepository : JpaRepository<House, UUID>
